nv-public: update bindings of tz & passive cdev
Before the userspace is up, the fan speed is governed based upon the
sensor reading of tj-thermal when kernel is up. However, when there is
userspace fan control service up and running, the policy of tj-thermal
shall be set to user_space so that kernel will no longer take the
control to avoid the conflict that could be introduced by the
multi-master.

Therefore, the passive cooling devices should be bound to
cpu/gpu/soc012/soc345 thermal zones which are functionally equivalent to
binding to tj-thermal zone. In this way, kernel will be only responsible
for passive cooling devices, and irresponsible for the fan speed control
unless the policy of tj-thermal is not set to user_space.

Signed-off-by: Yi-Wei Wang <yiweiw@nvidia.com>

Bug 5229600

Change-Id: I89ff96392f7b42b766a0e9f5d1690baa21dacbd8
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3348739
Reviewed-by: Yi-Wei Wang <yiweiw@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Reviewed-by: Rajkumar Kasirajan <rkasirajan@nvidia.com>
Tested-by: Yi-Wei Wang <yiweiw@nvidia.com>