t264: nv-platform: p3740: remove the key-suspend
The I/O PU.02 connects to ball L11 of both P3834 and P3740
connector. On P3834 the signal is named as STANDBY_REQ_N_GP172
and as STANDBY_REQ_N_GP109 on P3740.

On P3740,the STANDBY_REQ_N_GP109 pin is connected to pin 10
of the 12-pin debug header. This pin is used in both ways as the
"force recovery" pin as well as the "standby request" pin.

The gpio-keys configuration to control this pin on key-suspend event
is not correct and needs to be removed.

Bug 5277686

Change-Id: I2983c6a653851458dec41437aa667f5f9049c74f
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3363429
(cherry picked from commit 09a859c2313ec938acd8adba647e8c864dd00c59)
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3386725
Reviewed-by: Ninad Malwade <nmalwade@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Tested-by: Ninad Malwade <nmalwade@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>