dts: t264: Enable arm-timer in SC7
Our supporting platform code does implement sequences to keep time in
the SC7-state.
Removing this line to indicate to the Kernel during boot that the ARM
timer will be accounting for suspend time, and so it does not need to
skip that time as it does today.

Bug 5118451

Change-Id: I24834337d7500e62db80d6dc8869bdf349091c73
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3402419
Tested-by: Ishan Shah <ishah@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Reviewed-by: Ishan Shah <ishah@nvidia.com>