nv-public: p4071: update runtime check for ina238
The existing runtime overlay marks the status of ina238 to "okay" only
when the ids = ">=4071-0000-TS2", which fails to mark it to "okay" for
the "4071-0000-RC1".

Since the ina238 is avaiable for all the p4071 carrier boards except
EB1 and TS1, this change enables the ina238 by default. When the
"4071-0000-EB1" and "4071-0000-TS1" is detected, ina238 is disabled.

Bug 5297753

Change-Id: I639135de6766854eb2f4af4f90ee290635d795c0
Signed-off-by: Yi-Wei Wang <yiweiw@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3419040
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>