iommu/arm-smmu-v3: Implement WAR for Tegra264 Errata T264-SMMU-3
If there is active traffic on a page while it is being invalidated, it
is possible that the intended invalidated entry is not properly evicted
from the SMMU TLB caches. Subsequent requests to the same virtual
address could get the older mapping and access the wrong location, which
can be a functional and a security problem.
Repeat the entire invalidation->sync sequence once for every
invalidation queued. The second set of invalidation + sync commands is
guaranteed to evict the entry out of all TLBs.

Bug 5448714
Bug 5150304

Change-Id: I40e891a1459f4c65e4e2e7f0224867d6c1be277a
(cherry-picked from commit 27c7fb19c438457aabe7d7b4812bf88778750017)
Signed-off-by: Ashish Mhetre <amhetre@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/3rdparty/canonical/linux-noble/+/3496080
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Reviewed-by: Pritesh Raithatha <praithatha@nvidia.com>