t264: overlay: remove soc compatible string
Jetson-IO is supported on kenari but not Jedha board. However,
the overlays are still being applied in Jedha which is causing
the board to crash when CSI camera is selected. Remove soc
compatible string - 'nvidia,p3834-0008', and  'nvidia,tegra264'
in the overlays as it should be platform dependent alone.

Bug 5497230

Change-Id: Ife06683d6d1a8e30ca3096fed0394674f81c36c1
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3445125
Reviewed-by: Dipen Patel <dipenp@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Reviewed-by: Hiteshkumar Patel <hiteshkumarg@nvidia.com>
Tested-by: Gautham Srinivasan <gauthams@nvidia.com>