t264: nv-public: introduce bwmgr device node
This device node represents a virtual device in the kernel
that would be used for exposing the control interface of EMC
clock floor frequency via linux devfreq framework.

EMC clock handle is needed to construct the OPP table, which
is used for devfreq framework to save the valid frequency
steps.

BPMP handle is needed to allow the driver to communicate with
the BPMP via mailbox infrastructure.

Bug 5483386
Bug 5196455

Change-Id: I431c3e4cba1563aa313bdd20107b1839f1e00f6e
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3453759
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>
Reviewed-by: Johnny Liu <johnliu@nvidia.com>
Tested-by: Johnny Liu <johnliu@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
Reviewed-by: Ishan Shah <ishah@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>