t264: nv-public: Align GPIO compatible with upstream
The compatible string for upstream GPIO controllers for Tegra do not
include 'main' for the main GPIO controller it is simply
'nvidia,tegra<soc>-gpio'. To permit the upstreamed GPIO driver to work
with the downstream device-tree, add the upstream compatible for the
main controller along with the downstream to maintain backward
compatibility.

Bug 5757603

Change-Id: Id1eec46bd9dc5db598602bbb95f29abd2f76a10d
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/device/hardware/nvidia/t264-public-dts/+/3515892
Reviewed-by: Brad Griffis <bgriffis@nvidia.com>
GVS: buildbot_gerritrpt <buildbot_gerritrpt@nvidia.com>